Update cache-from in metadata step of build-and-test-job workflow
In the .github/workflows/build-and-test-job.yml file, the 'cache-from' argument has been updated in the build process in both stages. This change aims to improve the cache handling for the build-and-test-job by specifying a more precise source for the cache. The 'cache-to' argument has been commented for now but might be reintroduced in a later stage.